Backend/requirements.txt
aiohttp>=3.9.0
aiosqlite>=0.19.0
telethon>=1.37.0
python-dotenv>=1.0.0
cryptography>=42.0.0
